Fix wrong type of ClientId

+++ src/types.rs
@@ -1,6 +1,6 @@
 use rust_decimal::Decimal;
 
-pub type ClientId = u32;
+pub type ClientId = u16;
 
 pub type TransactionId = u32;
